Adding default_mode and current_season_mode switches
