+++ requirements.txt
@@ -1,10 +1,10 @@
-requests==2.25.1
+requests==2.26.0
 certifi==2020.12.5
 cffi==1.14.5
 chardet==4.0.0
-cryptography==3.3.2
+cryptography
 idna==2.10
-peewee==3.14.4
+peewee==3.14.8
 pycparser==2.20
 PyMySQL==1.0.0
 python-dotenv==0.15.0
