+++ scripts/create_postgres.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-DB_NAME="nba" DB_HOST="localhost" DB_USER=nba_sql DB_PASSWORD=nba_sql python stats/nba_sql.py --database="postgres" --create-schema
+DB_NAME="nba" DB_HOST="localhost" DB_USER=nba_sql DB_PASSWORD=nba_sql python stats/nba_sql.py --default_mode --database="postgres" --create-schema --skip-tables play_by_play pgtt player_season
