+++ scripts/create_sqlite.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-DB_NAME="nba" DB_HOST="localhost" DB_USER=nba_sql DB_PASSWORD=nba_sql python stats/nba_sql.py --database="sqlite" --create-schema --skip-tables play_by_play
+DB_NAME="nba" DB_HOST="localhost" DB_USER=nba_sql DB_PASSWORD=nba_sql python stats/nba_sql.py --default_mode --skip-tables play_by_play pgtt
